Include all the files in the press folder
Previously it was just PNGs, but the upload instructions indicate that links
to YouTube videos should be in text files, for example. This format should
be more inclusive of those at the risk of perhaps including unwanted
project files such as xcf. We will have to watch and exclude things we
don't want.